### 2021-10-07 #3805 @whedon
[PRE REVIEW]: Systole: A python package for cardiac signal synchrony and analysis
**Submitting author:** @LegrandNico (<a href="http://orcid.org/0000-0002-2651-4929">Nicolas Legrand</a>)
**Repository:** <a href="https://github.com/embodied-computation-group/systole/" target ="_blank">https://github.com/embodied-computation-group/systole/</a>
**Version:** v0.2.0
