### 2021-10-09 #3805 @osorensen
@LegrandNico, before we start the review, I also suggest you fix the following small issues iin the paper:

- Please make sure Python is spelled consistently. Currently both 'python' and 'Python' are being used.
- Lines 23-24: The sentence 'Similarly, rising interest in understanding how heart-rate variability relates to ...' is not completely. You can, e.g., rewrite to 'Similarly, there is a rising interesting in understanding ...'.
