### 2021-10-23 286d651
[JOSS review] Add tqdm to requirements
Installing this in a fresh python environment, I encountered this missing dependency. I didn't pin a version as I have not tried which versions work and which don't, and the tqdm API has been fairly stable for a while.
