+++ requirements.txt
@@ -7,3 +7,4 @@ pyserial>=3.4
 setuptools>=38.4
 packaging
 numba>=0.51.2
+tqdm
