### 2021-10-23 2f71224
[JOSS review] Fix typo in README.rst
Stepping through the examples in the readme, I encountered a missing `import` :)
