### 2021-10-28 6510b2a
Update requirements.txt to fix numba dependencies #19
