+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy>=1.17
+numpy>=1.17,<1.21
 scipy>=1.3.0
 pandas>=0.24
 matplotlib>=3.0.2
