### 2021-10-28 73c0394
Add bokeh in the requirements.txt as it is required during import, even if not used (#21).
