+++ requirements.txt
@@ -3,6 +3,7 @@ scipy>=1.3.0
 pandas>=0.24
 matplotlib>=3.0.2
 seaborn>=0.9.0
+bokeh>=2.3.2
 pyserial>=3.4
 setuptools>=38.4
 packaging
