### 2021-10-28 9b28457
[JOSS review] Fix link for lab logos (website)
