+++ source/index.rst
@@ -150,13 +150,13 @@ Contributors
 
 |AU| |lundbeck| |lab|
 
-.. |AU| image::  https://github.com/embodied-computation-group/systole/raw/dev/Images/au_clinisk_logo.png
+.. |AU| image::  https://github.com/embodied-computation-group/systole/blob/dev/source/images/au_clinisk_logo.png
    :width: 100%
 
-.. |lundbeck| image::  https://github.com/embodied-computation-group/systole/raw/dev/Images/lundbeckfonden_logo.png
+.. |lundbeck| image::  https://github.com/embodied-computation-group/systole/blob/dev/source/images/lundbeckfonden_logo.png
    :width: 10%
 
-.. |lab| image::  https://github.com/embodied-computation-group/systole/raw/dev/Images/LabLogo.png
+.. |lab| image::  https://github.com/embodied-computation-group/systole/blob/dev/source/images/LabLogo.png
    :width: 20%
 
 
