### 2021-10-28 0aa4d2c
[JOSS review] Fix citation format
