### 2021-11-07 #3832 @janfreyberg
All software-related review items look great to me now - the software installs and runs as documented. Pending the state of the field addition to the paper, all looks good to me :)

One point: I was unable to run the code related to online signal processing, as it requires hardware. This code is also not tested automatically. However, given the nature of the code, I don't think it's possible to change that without extensive mocking, which I think would be pointless, so happy to leave this untested!
