### 2021-11-07 #3832 @axel-loewe
I had to install `requests` as an additional dependency that was not listed.

Also, I had to add a couple of imports in order to run the Getting Started examples (e.g., `from systole.plots plot_raw`)
