### 2021-11-08 88d7de6
Add Gitter chat link in README
