+++ README.rst
@@ -26,6 +26,9 @@
 .. image:: https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white
   :target: https://github.com/pre-commit/pre-commit
 
+.. image:: https://badges.gitter.im/ecg-systole/community.svg
+   :target: https://gitter.im/ecg-systole/community?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge
+
 ================
 
 .. figure::  https://github.com/embodied-computation-group/systole/blob/dev/source/images/logo.png
@@ -38,8 +41,9 @@ This includes tools for data epoching, artefact detection, artefact correction, 
 variability analyses, circular statistical approaches to analysing cardiac cycles, and synchronising stimulus 
 presentation with different cardiac phases via Psychopy.
 
-
 The documentation can be found under the following `link <https://systole-docs.github.io/>`_.
+
+If you have questions, you can ask them in the `Gitter chat <https://gitter.im/ecg-systole/community?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge>`_.
 
 Installation
 ++++++++++++
