### 2021-11-08 #3832 @LegrandNico
* All import and documentation issues should be fixed now.

* We have created a public Gitter channel.

* l.43 The reference `@lipponen:2019` is not rendering correctly, even though, from what I can see, it follows the example provided on [JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html).

* Regarding the inspiration from other toolboxes, Systole is primarily made from code that was developed in the lab from scratch to implement new or preexisting algorithms, in which case the reference to the original article is mentioned in the docstring. Most of the time, no Python code was copied/duplicated. One main exception is the re-implementation of the ECG peaks detection algorithms from the py-ecg-detectors package. This package was previously part of the dependencies but we finally re-implemented the code directly in Systole to allow acceleration using Numba. We try to mention and credit this as much as possible in the docstrings, tutorial and example. It could maybe be more emphasised in the paper? Also, while we want to make this import explicit, it is also worth noting that it is not an identical copy, but that the functions have been significantly refactored and performance improved.

* We can run some comparisons with other software, the most straightforward one would be performance comparison on feature detection and artefact correction with other software (e.g. mne, neurokit). Is there a specific analysis/dataset we should be used in particular?
