### 2021-11-08 7d44594
Add Github action for testing
