### 2021-11-08 c2b60f9
Do not test for python <=3.6 and add pytest import
