### 2021-11-08 f52432a
Github actions: load requirements-test.txt by default
