### 2021-11-08 b672868
Github actions: installing systole
