### 2021-11-08 a74d538
Github actions: try to fix notebooks testing
