### 2021-11-08 f3ba2ed
Github actions: do not test notebooks
