### 2021-11-08 645f805
Github actions: install ipykernel
