### 2021-11-16 4e86756
[Fix] plot_evoked (plot_evoked does not work as expected for epochs inputs #26), so it accpets the three kind of inputs and improve the docstring.
