+++ systole/plots/plot_evoked.py
@@ -114,12 +114,19 @@ def plot_evoked(
     Examples
     --------
 
-    Plot evoked heart rate across two conditions using the Matplotlib backend.
+    Plot evoked heart rate across two conditions using the Matplotlib backend. Here,
+    for the sake of example, we are going to create the same plot three time using
+    three kind of input data:
+
+    * The raw signal + the triggers timing (or a list of in case of multiple conditions).
+    * The peaks detection + the triggers timing (or a list of in case of multiple conditions)
+    * The epoched signal as a 2d NumPy array (or a list of in case of multiple conditions)
 
     .. jupyter-execute::
 
        import numpy as np
        import seaborn as sns
+       import matplotlib.pyplot as plt
        from systole.detection import ecg_peaks
        from systole.plots import plot_evoked
        from systole import import_dataset1
@@ -135,26 +142,82 @@ def plot_evoked(
        # Peak detection in the ECG signal using the Pan-Tompkins method
        signal, peaks = ecg_peaks(ecg_df.ecg, method='pan-tompkins', sfreq=1000)
 
-       plot_evoked(
-           rr=peaks, triggers_idx=triggers_idx, input_type="peaks",
-           palette=[sns.xkcd_rgb["pale red"], sns.xkcd_rgb["denim blue"]],
-           apply_baseline=(-1.0, 0.0), decim=100, figsize=(400, 400)
-       )
+       # Convert to instantaneous heart rate
+       rr, _ = heart_rate(peaks, kind="cubic", unit="bpm", input_type="peaks")
 
-    Plot evoked heart rate across two conditions using the Bokeh backend.
+       # Create list epochs arrays for each condition
+       epochs, _ = to_epochs(
+           signal=rr, triggers_idx=triggers_idx, tmin=-1.0, tmax=10.0,
+           apply_baseline=(-1.0, 0.0)
+           )
+
+       fig, axs = plt.subplots(ncols=3, figsize=(15, 5), sharey=True)
+
+       # Using the raw signal and events triggers
+       plot_evoked(
+            signal=ecg_df.ecg.to_numpy(), triggers_idx=triggers_idx, modality="ecg",
+            tmin=-1.0, tmax=10.0, apply_baseline=(-1.0, 0.0), backend="matplotlib",
+            palette=[sns.xkcd_rgb["denim blue"], sns.xkcd_rgb["pale red"]],
+            ax=axs[0]
+            )
+
+       # Using the detected peaks and events triggers
+       plot_evoked(
+           rr=peaks, triggers_idx=triggers_idx, input_type="peaks", tmin=-1.0,
+           tmax=10.0, apply_baseline=(-1.0, 0.0), backend="matplotlib",
+           palette=[sns.xkcd_rgb["denim blue"], sns.xkcd_rgb["pale red"]],
+           ax=axs[1]
+           )
+
+       # Using the list of epochs arrays
+       plot_evoked(
+           epochs=epochs,
+           palette=[sns.xkcd_rgb["denim blue"], sns.xkcd_rgb["pale red"]],
+           backend="matplotlib",
+           ax=axs[2]
+           )
+
+    Plot evoked heart rate across two conditions using the Bokeh backend. Here,
+    for the sake of example, we are going to create the same plot three times using
+    three kind of input data:
+
+    * The raw signal + the triggers timing (or a list of in case of multiple conditions).
+    * The peaks detection + the triggers timing (or a list of in case of multiple conditions)
+    * The epoched signal as a 2d NumPy array (or a list of in case of multiple conditions)
 
     .. jupyter-execute::
 
        from bokeh.io import output_notebook
[…]
