+++ source/releases/v0.2.1.txt
@@ -1,0 +1,11 @@
+v0.2.1 (under progress)
+-----------------------
+
+Highlights
+++++++++++
+
+
+Changes
++++++++
+
+- |Fix| |Enhancement| Fix :func:`systole.plots.plot_evoked()` so it works with the three inputs types and improve documentation.
