### 2021-11-24 77f23a4
Use Numba's jit for norm_triggers
