### 2021-11-24 #3832 @LegrandNico
I have added a [sentence](https://github.com/embodied-computation-group/systole/commit/1777819ffd13462f88acccddac7aeba7b445802a) in the manuscript to clarify the unique features of the package.

If this one is accepted, I think I have fixed all the issues that were raised by the reviewers.

Just let me know if I can do anything to improve the package/manuscript.
