### 2021-11-24 embodied-computation-group/systole#27 @osorensen
Editorial comments to JOSS submission
Hello @LegrandNico. I'm the editor of your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/3832), and will create some issues for editorial comments to your paper.

On lines 22-24, you write

> Similarly, rising interest in understanding how heart-rate variability relates to mental illness, cognitive function, and physical wellbeing.

This is an incomplete sentence. Suggested rewrite: "Similarly, there is rising interest in understanding how heart-rate variability relates to mental illness, cognitive function, and physical wellbeing."
