### 2021-11-24 embodied-computation-group/systole#28 @osorensen
Inconsistent spelling of Python
On line 28 you write "python", while in the rest of the paper you write "Python". For consistency I suggest using "Python" throughout.

https://github.com/openjournals/joss-reviews/issues/3832
