### 2021-11-24 embodied-computation-group/systole#29 @osorensen
Package vs library
On line 8 you write "Systole is a library for cardiac signal analysis in Python.", whereas later, e.g., on line 33 you write "package". I think "package" is the most appropriate term to use here, so I suggest rewriting from "library" to "package" on line 8.

https://github.com/openjournals/joss-reviews/issues/3832
