### 2021-11-24 embodied-computation-group/systole#30 @osorensen
Reference not showing up properly
On line 46, a reference shows up as "`@lipponen:2019`". Please make sure this is corrected.

https://github.com/openjournals/joss-reviews/issues/3832
