### 2021-11-24 embodied-computation-group/systole#31 @osorensen
Bibliography
In the References section, lines 74-75, make sure the title appears as "Numba: a LLVM-based Python JIT compiler".

https://github.com/openjournals/joss-reviews/issues/3832
