### 2021-11-26 1a045a5
Fix Numba citation (#31)
