+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{numba,
-  title={Numba: A llvm-based python jit compiler},
+  title={Numba: a LLVM-based Python JIT compiler},
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
