### 2021-11-26 embodied-computation-group/systole#30 @LegrandNico
I have some difficulties understanding what is going on here, as from what I can see in [the JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html) the citation conforms to the syntax:

```
For a quick reference, the following citation commands can be used:
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
```
