### 2021-11-26 fb36823
[paper] : change "library" to "package" (#29)
