### 2021-11-26 667b8b5
[paper] : Use Python consistently (#28)
