### 2021-11-26 35e90c4
[paper] : Correct incomplete sentence (#27)
