### 2021-11-26 embodied-computation-group/systole#30 @osorensen
Agree that it's confusing, but the outer quotation marks should not be included. I checked that it compiles properly with this fixed, and created a PR.
