### 2021-11-29 d0a36fd
hrv: Important BUGFIX. the frequency_domain function switched hf_powernu and lf_power nu before returning. Also rename the percent and normalized variables to make it consistent with the other ones.
