### 2021-11-29 be0d50f
tests: Add exception for BrainVision in case of ConnectionRefusedError for some GitHub actions
