### 2021-11-29 1b10fdd
plot_frequency: clip power by default (>=0)
