+++ source/releases/v0.2.1.txt
@@ -12,4 +12,5 @@ Changes
 - |Fix| Important BUGFIX: :func:`systole.hrv.frequency_domain()` switched the variables `"power_hf_nu"` and `"power_lf_nu"` before returning.
 - |Fix| |Enhancement| Fix :func:`systole.plots.plot_evoked()` so it works with the three inputs types and improve documentation.
 - |Enhancement| Fix :func:`systole.utils.norm_triggers()` is now using Numba's jit by default.
-- |Enhancement| The :func:`systole.plot.plot_frequency()` function clip power to be >= 0 by default.
+- |Enhancement| |Defaults| The :func:`systole.plot.plot_frequency()` function clip power to be >= 0 by default.
+- |Defaults| The :func:`systole.hrv.frequency_domain()` :func:`systole.hrv.time_domain()` functions rounds with the same parameter (`6`).
