### 2021-11-29 c00648e
hrv: Fix normalized units (x100)
