### 2021-12-01 828dda7
plot_subspaces: improve docstring and add checks for input data
