### 2021-12-02 2e2b754
README: add table for tutorials
