### 2021-12-02 3aa8783
README: add table for tutorials (2)
