### 2021-12-02 076dfca
Fix image link in documnetation
