### 2021-12-02 eb30f7f
Add changes details in docs
