### 2021-12-02 c1357ce
Switch to v0.2.1 and change package title to match with the paper
