+++ setup.py
@@ -15,12 +15,12 @@ def get_requirements():
         return buff.read().splitlines()
 
 
-DESCRIPTION = """Cardiac signal analysis with Python"""
+DESCRIPTION = """Systole: A python package for cardiac signal synchrony and analysis"""
 
 DISTNAME = "systole"
 MAINTAINER = "Nicolas Legrand"
 MAINTAINER_EMAIL = "nicolas.legrand@cfin.au.dk"
-VERSION = "0.2.0"
+VERSION = "0.2.1"
 
 
 if __name__ == "__main__":
