+++ source/releases/v0.2.1.txt
@@ -1,5 +1,5 @@
-v0.2.1 (under progress)
------------------------
+v0.2.1 - December 2021
+----------------------
 
 Highlights
 ++++++++++
