### 2021-12-02 170faaa
plot_evoked: Fix import in docstring
