### 2021-12-02 3933f43
nonlinear_domain: - rename systole.hrv.nonlinear to systole.hrv.nonlinear_domain for consistency - Fix typo (pointcare) everywhere - Create systole.hrv.poincare function (with Numba JIT backend)
