+++ systole/hrv.py
@@ -4,19 +4,22 @@ from typing import Dict, List, Optional, Tuple, Union
 
 import numpy as np
 import pandas as pd
+from numba import jit
 from scipy import interpolate
 from scipy.signal import welch
 
 from systole.utils import input_conversion
 
 
-def nnX(x: Union[List, np.ndarray], t: int = 50, input_type: str = "rr_ms") -> float:
+def nnX(rr: Union[List, np.ndarray], t: int = 50, input_type: str = "rr_ms") -> float:
     """Number of difference in successive R-R interval > t ms.
 
     Parameters
     ----------
-    x : np.ndarray | list
-        Interval time-series (R-R in seconds or miliseconds, peaks or peaks indexes).
+    rr : np.ndarray | list
+        R-R interval time-series, peaks or peaks index vectors. The default expected
+        vector is R-R intervals in milliseconds. Other data format can be provided by
+        specifying the `"input_type"` (can be `"rr_s"`, `"peaks"` or `"peaks_idx"`).
     t : int
         Threshold value: Defaut is set to 50 ms to calculate the nn50 index.
     input_type : str
@@ -30,26 +33,29 @@ def nnX(x: Union[List, np.ndarray], t: int = 50, input_type: str = "rr_ms") -> f
 
     """
 
-    x = np.asarray(x)
+    rr = np.asarray(rr)
 
     if input_type != "rr_ms":
-        x = input_conversion(x, input_type=input_type, output_type="rr_ms")
+        rr = input_conversion(rr, input_type=input_type, output_type="rr_ms")
 
-    if len(x.shape) > 1:
+    if len(rr.shape) > 1:
         raise ValueError("X must be a 1darray")
 
     # NN50: number of successive differences larger than t ms
-    nn = np.sum(np.abs(np.diff(x)) > t)
+    nn = np.sum(np.abs(np.diff(rr)) > t)
+
     return nn
 
 
-def pnnX(x: Union[List, np.ndarray], t: int = 50, input_type: str = "rr_ms") -> float:
+def pnnX(rr: Union[List, np.ndarray], t: int = 50, input_type: str = "rr_ms") -> float:
     """Number of successive differences larger than a value (def = 50ms).
 
     Parameters
     ----------
-    x : np.ndarray | list
-        Interval time-series (R-R in seconds or miliseconds, peaks or peaks indexes).
+    rr : np.ndarray | list
+        R-R interval time-series, peaks or peaks index vectors. The default expected
+        vector is R-R intervals in milliseconds. Other data format can be provided by
+        specifying the `"input_type"` (can be `"rr_s"`, `"peaks"` or `"peaks_idx"`).
     t : int
         Threshold value: Defaut is set to 50 ms to calculate the nn50 index.
     input_type : str
@@ -63,30 +69,32 @@ def pnnX(x: Union[List, np.ndarray], t: int = 50, input_type: str = "rr_ms") -> 
 
     """
 
-    x = np.asarray(x)
+    rr = np.asarray(rr)
 
     if input_type != "rr_ms":
-        x = input_conversion(x, input_type=input_type, output_type="rr_ms")
+        rr = input_conversion(rr, input_type=input_type, output_type="rr_ms")
 
-    if len(x.shape) > 1:
+    if len(rr.shape) > 1:
         raise ValueError("X must be a 1darray")
 
     # nnX: number of successive differences larger than t ms
-    nn = nnX(x, t)
+    nn = nnX(rr, t)
 
     # Proportion of successive differences larger than t ms
-    pnnX = 100 * nn / len(np.diff(x))
[…]
