+++ systole/hrv.py
@@ -142,22 +142,23 @@ def time_domain(rr: Union[List, np.ndarray], input_type: str = "rr_ms") -> pd.Da
     -------
     stats : :py:class:`pandas.DataFrame`
         Time domain summary statistics.
-        - ``'Mean RR'`` : Mean of R-R intervals.
-        - ``'Mean BPM'`` : Mean of beats per minutes.
-        - ``'Median RR'`` : Median of R-R intervals'.
-        - ``'Median BPM'`` : Meidan of beats per minutes.
-        - ``'MinRR'`` : Minimum R-R intervals.
-        - ``'MinBPM'`` : Minimum beats per minutes.
-        - ``'MaxRR'`` : Maximum R-R intervals.
-        - ``'MaxBPM'`` : Maximum beats per minutes.
-        - ``'SDNN'`` : Standard deviation of successive differences.
-        - ``'RMSSD'`` : Root Mean Square of the Successive Differences.
-        - ``'NN50'`` : number of successive differences larger than 50ms.
-        - ``'pNN50'`` : Proportion of successive difference larger than 50ms.
+        - `'Mean RR'` : Mean of R-R intervals.
+        - `'Mean BPM'` : Mean of beats per minutes.
+        - `'Median RR'` : Median of R-R intervals'.
+        - `'Median BPM'` : Meidan of beats per minutes.
+        - `'MinRR'` : Minimum R-R intervals.
+        - `'MinBPM'` : Minimum beats per minutes.
+        - `'MaxRR'` : Maximum R-R intervals.
+        - `'MaxBPM'` : Maximum beats per minutes.
+        - `'SDNN'` : Standard deviation of RR intervals.
+        - `'SDSD'`: Standard deviation of the Successive difference.
+        - `'RMSSD'` : Root Mean Square of the Successive Differences.
+        - `'NN50'` : number of successive differences larger than 50ms.
+        - `'pNN50'` : Proportion of successive difference larger than 50ms.
 
     See also
     --------
-    frequency_domain, nonlinear
+    frequency_domain, nonlinear_domain
 
     Notes
     -----
@@ -208,6 +209,9 @@ def time_domain(rr: Union[List, np.ndarray], input_type: str = "rr_ms") -> pd.Da
     # Standard deviation of R-R intervals
     sdnn = round(rr.std(ddof=1), 6)  # type: ignore
 
+    # Standard deviation of the difference of successive R-R intervals
+    sdsd = round(np.diff(rr).std(ddof=1), 6)  # type: ignore
+
     # Root Mean Square of Successive Differences (RMSSD)
     rms = round(rmssd(rr), 6)
 
@@ -228,6 +232,7 @@ def time_domain(rr: Union[List, np.ndarray], input_type: str = "rr_ms") -> pd.Da
         max_rr,
         max_bpm,
         sdnn,
+        sdsd,
         rms,
         nn,
         pnn,
@@ -242,6 +247,7 @@ def time_domain(rr: Union[List, np.ndarray], input_type: str = "rr_ms") -> pd.Da
         "MaxRR",
         "MaxBPM",
         "SDNN",
+        "SDSD",
         "RMSSD",
         "nn50",
         "pnn50",
@@ -267,8 +273,7 @@ def psd(
         vector is R-R intervals in milliseconds. Other data format can be provided by
         specifying the `"input_type"` (can be `"rr_s"`, `"peaks"` or `"peaks_idx"`).
     sfreq : int
-        The sampling frequency (Hz) of the interpolated instantaneous heart
-        rate.
+        The sampling frequency (Hz) of the interpolated instantaneous heart rate.
     method : str
[…]
