+++ source/changelog.rst
@@ -24,6 +24,8 @@ This page contains information about what has changed in each new version of ``s
 
    <div class="col-md-9">
 
+.. include:: releases/v0.2.1.txt
+
 .. include:: releases/v0.2.0.txt
 
 .. include:: releases/v0.1.3.txt
