+++ systole/hrv.py
@@ -593,7 +593,7 @@ def _poincare(rr: np.ndarray) -> Tuple[float, float]:
 def recurrence(
     rr: Union[List, np.ndarray], input_type: str = "rr_ms"
 ) -> Tuple[float, int, float, float, float]:
-    """Compute SD1 and SD2 from the recurrence plot [1]_ for heart rate variability.
+    """Compute quantitative metrics from the recurrence plot for heart rate variability.
 
     Parameters
     ----------
@@ -608,13 +608,14 @@ def recurrence(
     Returns
     -------
     recurrence_rate : float
-        Recurence rate. The ratio of ones and zeros in the recurrence plot.
+        The percentage of recurence in the time series. This corresponds to the ratio
+        of ones and zeros in the recurrence plot.
     l_max : int
         Maximum lenght of the diagonale in the reccurence plot.
     l_mean : float
         Mean of the diagonals lengths observed in the recurence plot.
-    determinism : float
-        Determinism of the time series.
+    determinism_rate : float
+        The percentage of determinism in the time series.
     shan_entr : float
         Shannon information entropy.
 
@@ -623,7 +624,7 @@ def recurrence(
 
     See also
     --------
-    nonlinear_domain
+    nonlinear_domain, poincare
 
     References
     ----------
@@ -635,55 +636,67 @@ def recurrence(
        systems and states using recurrence plot strategies. J Appl Physiol, 76:965–973,
        1994.
 
+    .. [3] Zbilut J. P., Webber C. L., Zak M.Quantification of heart rate variability
+       using methods derived from nonlinear dynamics.Assessment and Analysis of
+       Cardiovascular Function, Drzewiecki G., Li J. K.-J. Springer New York.
+
     """
     rr = np.asarray(rr)
 
     if input_type != "rr_ms":
         rr = input_conversion(rr, input_type=input_type, output_type="rr_ms")
 
-    recurrence_rate, l_max, l_mean, determinism, shan_entr = _recurrence(rr)
+    recurrence_rate, l_max, l_mean, determinism_rate, shan_entr = _recurrence(rr)
 
-    return recurrence_rate, l_max, l_mean, determinism, shan_entr
+    return recurrence_rate, l_max, l_mean, determinism_rate, shan_entr
 
 
-def _recurrence(rr: np.array, m: int = 10):
+def _recurrence(rr: np.array, m: int = 10, l_min: int = 2):
     """Compute recurrence scores"""
 
     # Recurrence matrix
     rc = recurrence_matrix(rr)
 
-    n = rc.shape[0]
-    recurrence_rate = np.sum(rc) / (n ** 2)
+    # Compute the recurrence rate - Exclude the main identity line
+    j = rc.shape[0]
+    recurrence_rate = np.triu(rc).sum() / ((j ** 2 - j) / 2) * 100
 
     # Find diagonale lines lines
     total_lines = []
-    for i in range(1, rc.shape[0]):
+    for i in range(1, rc.shape[0] // 2):
 
         # All diagonals except the main one
         diag = np.diagonal(rc, offset=i)
 
-        # Lenght of each diagonale with consecutive True values
+        # Lenght of each diagonale found with consecutive `True` values
         d = np.diff(
             np.where(np.concatenate(([diag[0]], diag[:-1] != diag[1:], [True])))[0]
         )[::2]
 
+        # Store the result if any
         if d.shape[0] > 0:
             total_lines.extend(d)
 
     # Compute scores
     l_max = np.max(total_lines)
-    l_lines = np.asarray(total_lines).repeat(2)
-    l_lines = l_lines[np.where(l_lines > 2)[0]]
 
[…]
