### 2021-12-04 5ce0ebe
- frequency_domain: Add the total power and the LF/HF ratio in the default output. - Fix various mypy errors
