+++ .pre-commit-config.yaml
@@ -5,7 +5,7 @@ repos:
     - id: isort
       files: ^systole/
 -   repo: https://github.com/ambv/black
-    rev: 21.9b0
+    rev: 21.11b1
     hooks:
     - id: black
       language_version: python3
@@ -16,7 +16,7 @@ repos:
     - id: flake8
       files: ^systole/
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: 'v0.910'  # Use the sha / tag you want to point at
+    rev: 'v0.910-1'  # Use the sha / tag you want to point at
     hooks:
     - id: mypy 
       files: ^systole/
