### 2021-12-04 4fa3361
- Add res_peaks function of peaks detection in respiratory signal (very simple for now). - Minor changes in docstrings
