### 2021-12-04 d31e587
Fix error in pan-tompkins (after Mypy fix)
