### 2021-12-04 a2476bf
hrv: faster implementation for recurrence matrix (#33)
