+++ source/releases/v0.2.2.txt
@@ -1,0 +1,19 @@
+v0.2.2 (December 2021 - 2)
+--------------------------
+
+Highlights
+++++++++++
+
+This release seeks to improve the heart rate variability metrics implemented in Systole and provide tools for interactive reports creation and group results summary from any BIDS structured folders containing physiological recording.
+
+Changes
++++++++
+
+- |Feature| HRV module: Add recurence plot analysis and fast recurrence matrix computation (thanks to Dominique Makowski), and basic quantitative metrics (recurrence rate, l_max, l_mean, determinism rate, Shannon entropy). These metrics are now part of the nonlinear reports (py:func:`systole.hrv.nonlinear_domain()`) or can be called separately (py:func:`systole.hrv.recurrence()`).
+- |Feature| HRV module: The time domain function (py:func:`systole.hrv.time_domain()`) also reports the SDSD by default.
+- |Feature| HRV module: The frequency domain function (py:func:`systole.hrv.frequency_domain()`) now returns the total power and LF/HF ratio by default.
+
+Contributors
+++++++++++++
+
+- Dominique Makowski (https://github.com/DominiqueMakowski)
