### 2021-12-04 ea69afd
Remove Travis testing from workflow
