+++ systole/plots/utils.py
@@ -1,8 +1,15 @@
 # Author: Nicolas Legrand <nicolas.legrand@cfin.au.dk>
 
 import importlib
+from typing import List, Optional, Union, overload
 
+import numpy as np
+import pandas as pd
+from bokeh.models import Column, ColumnDataSource, DataTable, TableColumn
 from packaging.version import parse
+from tabulate import tabulate
+
+from systole.hrv import frequency_domain, nonlinear_domain, time_domain
 
 
 def get_plotting_function(plot_name, plot_module, backend="matplotlib"):
@@ -50,3 +57,400 @@ def get_plotting_function(plot_name, plot_module, backend="matplotlib"):
     plotting_method = getattr(module, plot_name)
 
     return plotting_method
+
+
+@overload
+def time_table(
+    rr: None,
+    input_type: None,
+    time_df=pd.DataFrame,
+) -> Union[Column, str]:
+    ...
+
+
+@overload
+def time_table(
+    rr: Union[List, np.ndarray],
+    input_type: str = "rr_ms",
+    time_df=None,
+) -> Union[Column, str]:
+    ...
+
+
+def time_table(
+    rr=None,
+    input_type="rr_ms",
+    time_df=None,
+    backend: str = "tabulate",
+    width: Optional[int] = 600,
+    height: Optional[int] = 300,
+) -> Union[Column, str]:
+    """Format time domain results for nice rendering.
+
+    Parameters
+    ----------
+    rr : np.ndarray | list
+        R-R interval time-series, peaks or peaks index vectors. The default expected
+        vector is R-R intervals in milliseconds. Other data format can be provided by
+        specifying the `"input_type"` (can be `"rr_s"`, `"peaks"` or `"peaks_idx"`).
+    input_type : str
+        The type of input provided. Can be `"peaks"`, `"peaks_idx"`, `"rr_ms"` or
+        `"rr_s"`. Defaults to `"rr_ms"`.
+    time_df : pd.DataFrame | None
+        The time domain results obtained from py:func:`systole.hrv.time_domain`.
+    bakend : str
+        Which backend to use. Can be `"bokeh"` or `"tabulate"`. Defaults to `"tabulate"`.
+    width, height : int
+        The table width and height (only for `"bokeh"` backend).
+
+    Returns
+    -------
+    table : `bokeh.models.Column` | str
+        The formatted time domain table, either as a string or as a Bokeh Columns.
+
+    """
+
+    if time_df is None:
+        if (rr is None) | (input_type is None):
+            raise ValueError(
+                "If no summary dataframe is provided, you must provide an RR time series."
+            )
+        time_df = time_domain(rr=rr, input_type=input_type)
+
+    data = {
+        "variable": [
+            "Mean RR",
+            "Median RR",
+            "Minimun RR",
+            "Maximum RR",
+            None,
+            "Mean BPM",
+            "Median BPM",
+            "Minimun BPM",
+            "Maximum BPM",
+            None,
+            "SDNN",
+            "SDSD",
+            "RMSSD",
+            "nn50",
+            "pnn50",
+        ],
+        "units": [
+            "milliseconds (ms)",
+            "milliseconds (ms)",
+            "milliseconds (ms)",
+            "milliseconds (ms)",
+            None,
+            "beats per minute (bpm)",
+            "beats per minute (bpm)",
+            "beats per minute (bpm)",
+            "beats per minute (bpm)",
+            None,
+            "milliseconds (ms)",
+            "milliseconds (ms)",
+            "milliseconds (ms)",
+            "counts",
+            "%",
+        ],
+        "value": [
[…]
