+++ systole/tests/test_plots.py
@@ -6,9 +6,11 @@ from unittest import TestCase
 import numpy as np
 import pandas as pd
 import seaborn as sns
+from bokeh.models import Column
 
 from systole import import_dataset1, import_ppg, import_rr
 from systole.detection import ecg_peaks, rr_artefacts
+from systole.hrv import frequency_domain, nonlinear_domain, time_domain
 from systole.plots import (
     plot_circular,
     plot_ectopic,
@@ -21,6 +23,7 @@ from systole.plots import (
     plot_shortlong,
     plot_subspaces,
 )
+from systole.plots.utils import frequency_table, nonlinear_table, time_table
 from systole.utils import heart_rate, to_epochs
 
 
@@ -199,6 +202,72 @@ class TestPlots(TestCase):
 
         with self.assertRaises(ValueError):
             plot_subspaces(rr=None, artefacts=None)
+
+    def test_time_table(self):
+        """Test the time_table function"""
+        rr = import_rr().rr
+        time_df = time_domain(rr, input_type="rr_ms")
+
+        # With a df as input
+        table_df = time_table(time_df=time_df, backend="tabulate")
+        assert isinstance(table_df, str)
+
+        table = time_table(time_df=time_df, backend="bokeh")
+        assert isinstance(table, Column)
+
+        # With RR intervals as inputs
+        table_rr = time_table(rr=rr, backend="tabulate")
+        assert isinstance(table_rr, str)
+
+        table = time_table(rr=rr, backend="bokeh")
+        assert isinstance(table, Column)
+
+        # Check for consistency between methods
+        assert table_rr == table_df
+
+    def test_frequency_table(self):
+        """Test plot_subspaces function"""
+        rr = import_rr().rr
+        frequency_df = frequency_domain(rr, input_type="rr_ms")
+
+        # With a df as input
+        table_df = frequency_table(frequency_df=frequency_df, backend="tabulate")
+        assert isinstance(table_df, str)
+
+        table = frequency_table(frequency_df=frequency_df, backend="bokeh")
+        assert isinstance(table, Column)
+
+        # With RR intervals as inputs
+        table_rr = frequency_table(rr=rr, backend="tabulate")
+        assert isinstance(table_rr, str)
+
+        table = frequency_table(rr=rr, backend="bokeh")
+        assert isinstance(table, Column)
+
+        # Check for consistency between methods
+        assert table_rr == table_df
+
+    def test_nonlinear_table(self):
+        """Test plot_subspaces function"""
+        rr = import_rr().rr
+        nonlinear_df = nonlinear_domain(rr, input_type="rr_ms")
+
+        # With a df as input
+        table_df = nonlinear_table(nonlinear_df=nonlinear_df, backend="tabulate")
+        assert isinstance(table_df, str)
+
+        table = nonlinear_table(nonlinear_df=nonlinear_df, backend="bokeh")
+        assert isinstance(table, Column)
+
+        # With RR intervals as inputs
+        table_rr = nonlinear_table(rr=rr, backend="tabulate")
+        assert isinstance(table_rr, str)
+
+        table = nonlinear_table(rr=rr, backend="bokeh")
+        assert isinstance(table, Column)
+
+        # Check for consistency between methods
+        assert table_rr == table_df
 
 
 if __name__ == "__main__":
