### 2021-12-16 ca9af23
[HRV] : - Add `all_domain` function to extract all possible indices in one function call. - Docstrings
