### 2021-12-16 c40c3d0
Documentation: Add missing function to API references
