### 2021-12-17 9bc6bda
Polish documentation and prepare for v0.2.2
