+++ source/api.rst
@@ -6,7 +6,7 @@ Functions
 =========
 
 .. contents:: Table of Contents
-   :depth: 3
+   :depth: 2
 
 Correction
 ----------
