+++ setup.py
@@ -1,6 +1,7 @@
 import os
 import codecs
 from setuptools import find_packages, setup
+from systole import __version__
 
 PROJECT_ROOT = os.path.dirname(os.path.realpath(__file__))
 REQUIREMENTS_FILE = os.path.join(PROJECT_ROOT, "requirements.txt")
@@ -20,7 +21,7 @@ DESCRIPTION = """Systole: A python package for cardiac signal synchrony and anal
 DISTNAME = "systole"
 MAINTAINER = "Nicolas Legrand"
 MAINTAINER_EMAIL = "nicolas.legrand@cfin.au.dk"
-VERSION = "0.2.2"
+VERSION = __version__
 
 
 if __name__ == "__main__":
