+++ systole/__init__.py
@@ -5,4 +5,4 @@ from .hrv import *
 from .plots import *
 from .utils import *
 
-__version__ = "0.2.0a"
+__version__ = "0.2.2"
