### 2021-12-17 a06658d
Fix image in documentation
