### 2021-12-17 #3832 @LegrandNico
I have just release v0.2.2. The list of version tags is the following:

* 0.0.1
* 0.1.0
* 0.1.1
* 0.1.2
* 0.1.3
* 0.2.0
* 0.2.1
* 0.2.2

The DOI for the last (0.2.2) Zenodo version is **10.5281/zenodo.5787520**

@micahgallen can you check the Zenodo metadata and let us know when this is ready?
