### 2021-12-24 #3832 @arfon
@LegrandNico – could you please make sure to cite Scipy, Numpy, Pandas and any other package you mention in your paper please?
