### 2022-01-05 #3832 @LegrandNico
I have added the missing citations for Numpy, SciPy and Pandas where they are mentioned in the summary. For Pandas I used the template provided on the website but I think the author list (`The pandas development team`) is not rendered correctly in the compiled paper.
