### 2022-01-05 #3832 @arfon
@LegrandNico – I made a slight tweak to your paper citation formatting in https://github.com/embodied-computation-group/systole/pull/36 .
